Fix leak of SSL caches for auxiliary threads
Whenever a thread creates MySQL connections using SSL params the
thread-local caches are populated. These thread-local resources need to
be free on thread exit via 'mysql_thread_end', otherwise these objects
will be leak. This is not relevant for the fixed size thread-pools, but
it's for workloads making use of temporary/auxiliary threads.